+++ client/cs_prefs.cpp
@@ -237,7 +237,7 @@ int CLIENT_STATE::check_suspend_processing() {
             return SUSPEND_REASON_TIME_OF_DAY;
         }
         if (global_prefs.suspend_if_no_recent_input) {
-            int idle_time = host_info.user_idle_time(check_all_logins);
+            long idle_time = host_info.user_idle_time(check_all_logins);
             if (idle_time != USER_IDLE_TIME_INF
                 && idle_time > global_prefs.suspend_if_no_recent_input*60
             ) {
